Improved binomial coefficient calculations
The old way used factorial which has a hard limit on the order of bezier
curves supported, due to the fact that sizeof(size_t) is easily exceeded
when calculation factorials.

For example, on systems where sizeof(size_t) == 4, only Bezier curves up
to order 12 was supported because factorial(13) > the maximum value of
size_t, which then overflowed.