tests: Use Catch2 for unit tests
Instead of using assert macros for testing it is better to use an
actual framework to run the unit tests. Catch2 is chosen because of its
simplicity and easiness to include and run through CMake.

This commit does not change any unit tests, but they are converted to
be tested using the Catch2 macros.